Fix SteamVR_UpdatePoses to work with every Unity version.
The `SteamVR_UpdatePoses` script used a dummy camera to be able to
update the poses using the `MonoBehaviour.OnPreCull` message method.
This fix subscribes to the static `Camera.onPreCull` event that is fired
before **any** camera starts culling.

This means the `SteamVR_UpdatePoses` script no longer needs to be on a
game object that has a `Camera` attached to it. Therefore the
requirement for the `Camera` component, as well as the setup of the
dummy camera is removed. To make sure the poses are only updated once
and not for **every** camera all cameras attached to game objects other
than `SteamVR_Render.Top().gameObject` are ignored.