ensure PlayArea vertices are consistent
The `vertices` variable in `SteamVR_PlayArea` is inconsistent when
setting the size between Calibrated and a fixed size.

The points of the vertices are in different locations based on whether
the size is either Calibrated or a given size which causes issues when
attempting to use the `vertices` variable to generate play area
representations.

The solution is to ensure the points of the fixed size match the
positions that a calibrated play area also has.