Fix disabled controllers when re-enabling SteamVR_ControllerManager.
The `SteamVR_ControllerManager` script set all game objects inactive in
its `OnEnable` method. If the Controller Manager is disabled and later
enabled again turned on controllers will not be able to be used in Unity
because their game object isn't kept active.

The fix is to make sure to only disable objects that have an invalid
index currently.