Allow Throwable extensions
The current Throwable implementation allows for subclassing but the use
of private methods prevents the subclass from overriding the methods
used to handle messages from Hand.cs. Changing the definition allows the
subclass the ability to override each method and still be able to call
the parent classes implementation as well. It's possible to use the onPickUp or onDetachFromHand but this can
be limiting.