fix: don't set worse fixedDeltaTime
SteamVR sets the fixedDeltaTime to a worse value than may be manually set in the Time settings in Unity due to HMD refresh being slower than the time specified. If the fixedDeltaTime set by SteamVR would be set to a worse value than is currently chosen, then leave it where it is.