Update Teleport.cs
Added the code to invoke Unity Events: 1) on specific TeleportPoints and TeleportAreas when the player telports to these and 2) on each teleport action (through the Teleporting prefab). This allows you to create events when the user teleports to a specific spot (like open a door in a dungeon) or allows you to run code whenever the player teleports (like increasing the turn in a turn based game or dynamically place new TeleportPoints on Unity terrain).