Teleport.cs: Check for existence of controllers before accessing them
When trying to work on a project without an HMD with controllers, the
project defaults to 2D Debug mode. However, since the code expects the
controllers to exist, it's not able to get past trying to execute haptic
feedback which results in teleporting to fail in 2D Debug mode. These
changes make sure that the script check for controller existence before
trying to access them.

This fixes issue #483